requirements.txt
feedparser
requests
beautifulsoup4
openai
load_dotenv